+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add mnemonic to Preferences menu in scaffolding
 - Make bin/glimmer, bin/girb, and scaffolded bin/script files call jruby instead of ruby in the shebang
 - Launch scaffolded app on Linux without using packaging (via `glimmer bin/app_script`)
-- Add both Mac and Windows icons upon scaffolding (not just the OS we are on)
+- Add all of Mac, Windows, and Linux icons upon scaffolding (not just for the OS we are on)
 
 **Packaging:**